Update performer cards
