+++ app/src/main/java/com/github/damontecres/stashapp/presenters/PerformerPresenter.kt
@@ -1,10 +1,14 @@
 package com.github.damontecres.stashapp.presenters
 
+import android.os.Build
 import androidx.leanback.widget.ImageCardView
-import androidx.preference.PreferenceManager
 import com.bumptech.glide.Glide
+import com.github.damontecres.stashapp.R
 import com.github.damontecres.stashapp.api.fragment.PerformerData
+import com.github.damontecres.stashapp.data.DataType
+import com.github.damontecres.stashapp.util.ageInYears
 import com.github.damontecres.stashapp.util.createGlideUrl
+import java.util.EnumMap
 
 class PerformerPresenter : StashPresenter() {
     override fun onBindViewHolder(
@@ -18,12 +22,19 @@ class PerformerPresenter : StashPresenter() {
             val title =
                 performer.name + (if (!performer.disambiguation.isNullOrBlank()) " (${performer.disambiguation})" else "")
             cardView.titleText = title
-            cardView.contentText = "${performer.scene_count} Scenes"
+
+            if (performer.birthdate != null && Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                val yearsOldStr = cardView.context.getString(R.string.stashapp_years_old)
+                cardView.contentText = "${performer.ageInYears} $yearsOldStr"
+            }
+
+            val dataTypeMap = EnumMap<DataType, Int>(DataType::class.java)
+            dataTypeMap[DataType.SCENE] = performer.scene_count
+
+            setUpExtraRow(cardView, dataTypeMap, performer.o_counter)
+
             cardView.setMainImageDimensions(CARD_WIDTH, CARD_HEIGHT)
-            val apiKey =
-                PreferenceManager.getDefaultSharedPreferences(vParent.context)
-                    .getString("stashApiKey", "")
-            val url = createGlideUrl(performer.image_path, apiKey)
+            val url = createGlideUrl(performer.image_path, cardView.context)
             Glide.with(viewHolder.view.context)
                 .load(url)
                 .centerCrop()
