+++ app/src/main/java/com/github/damontecres/stashapp/presenters/ScenePresenter.kt
@@ -5,16 +5,11 @@ import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.graphics.Color
 import android.graphics.Paint
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
 import androidx.leanback.widget.ImageCardView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.bitmap_recycle.BitmapPool
 import com.bumptech.glide.load.resource.bitmap.BitmapTransformation
 import com.bumptech.glide.load.resource.bitmap.CenterCrop
-import com.github.damontecres.stashapp.R
 import com.github.damontecres.stashapp.api.fragment.SlimSceneData
 import com.github.damontecres.stashapp.data.DataType
 import com.github.damontecres.stashapp.util.Constants
@@ -22,6 +17,7 @@ import com.github.damontecres.stashapp.util.concatIfNotBlank
 import com.github.damontecres.stashapp.util.createGlideUrl
 import com.github.damontecres.stashapp.util.titleOrFilename
 import java.security.MessageDigest
+import java.util.EnumMap
 
 class ScenePresenter : StashPresenter() {
     override fun onBindViewHolder(
@@ -38,16 +34,13 @@ class ScenePresenter : StashPresenter() {
         details.add(scene.date)
         cardView.contentText = concatIfNotBlank(" - ", details)
 
-        val infoView = cardView.findViewById<ViewGroup>(androidx.leanback.R.id.info_field)
-        val sceneExtra =
-            LayoutInflater.from(infoView.context)
-                .inflate(R.layout.scene_card_extra, infoView, true) as ViewGroup
+        val dataTypeMap = EnumMap<DataType, Int>(DataType::class.java)
+        dataTypeMap[DataType.TAG] = scene.tags.size
+        dataTypeMap[DataType.PERFORMER] = scene.performers.size
+        dataTypeMap[DataType.MOVIE] = scene.movies.size
+        dataTypeMap[DataType.MARKER] = scene.scene_markers.size
 
-        setUpIcon(sceneExtra, DataType.TAG, scene.tags.size)
-        setUpIcon(sceneExtra, DataType.PERFORMER, scene.performers.size)
-        setUpIcon(sceneExtra, DataType.MOVIE, scene.movies.size)
-        setUpIcon(sceneExtra, DataType.MARKER, scene.scene_markers.size)
-        setUpIcon(sceneExtra, null, scene.o_counter ?: -1)
+        setUpExtraRow(cardView, dataTypeMap, scene.o_counter)
 
         if (!scene.paths.screenshot.isNullOrBlank()) {
             cardView.setMainImageDimensions(CARD_WIDTH, CARD_HEIGHT)
@@ -57,55 +50,6 @@ class ScenePresenter : StashPresenter() {
                 .transform(CenterCrop(), TextOverlay(viewHolder.view.context, scene))
                 .error(mDefaultCardImage)
                 .into(cardView.mainImageView!!)
-        }
-    }
-
-    private fun setUpIcon(
-        rootView: ViewGroup,
-        dataType: DataType?,
-        count: Int,
-    ) {
-        val textResId: Int
-        val iconResId: Int
-        when (dataType) {
-            DataType.MOVIE -> {
-                textResId = R.id.scene_movie_count
-                iconResId = R.id.scene_movie_icon
-            }
-
-            DataType.MARKER -> {
-                textResId = R.id.scene_marker_count
-                iconResId = R.id.scene_marker_icon
-            }
-
-            DataType.PERFORMER -> {
-                textResId = R.id.scene_performer_count
-                iconResId = R.id.scene_performer_icon
-            }
-
-            DataType.TAG -> {
-                textResId = R.id.scene_tag_count
-                iconResId = R.id.scene_tag_icon
-            }
-
-            // Workaround for O Counter
-            null -> {
-                textResId = R.id.scene_ocounter_count
-                iconResId = R.id.scene_ocounter_icon
-            }
-
-            else -> throw IllegalArgumentException()
-        }
-        val textView = rootView.findViewById<TextView>(textResId)
-        val iconView = rootView.findViewById<View>(iconResId)
-        if (count > 0) {
-            textView.text = count.toString()
-            textView.visibility = View.VISIBLE
-            iconView.visibility = View.VISIBLE
-        } else {
-            textView.visibility = View.GONE
-            iconView.visibility = View.GONE
-            (textView.parent as ViewGroup).visibility = View.GONE
         }
     }
