+++ app/src/main/java/com/github/damontecres/stashapp/presenters/StashPresenter.kt
@@ -1,6 +1,8 @@
 package com.github.damontecres.stashapp.presenters
 
 import android.graphics.drawable.Drawable
+import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.core.content.ContextCompat
@@ -14,11 +16,13 @@ import com.github.damontecres.stashapp.api.fragment.MovieData
 import com.github.damontecres.stashapp.api.fragment.PerformerData
 import com.github.damontecres.stashapp.api.fragment.SlimSceneData
 import com.github.damontecres.stashapp.api.fragment.StudioData
+import com.github.damontecres.stashapp.data.DataType
 import com.github.damontecres.stashapp.data.OCounter
 import com.github.damontecres.stashapp.data.StashCustomFilter
 import com.github.damontecres.stashapp.data.StashSavedFilter
 import com.github.damontecres.stashapp.data.Tag
 import com.github.damontecres.stashapp.util.enableMarquee
+import java.util.EnumMap
 import kotlin.properties.Delegates
 
 abstract class StashPresenter : Presenter() {
@@ -77,6 +81,75 @@ abstract class StashPresenter : Presenter() {
         // during animations.
         view.setBackgroundColor(color)
         view.setInfoAreaBackgroundColor(color)
+    }
+
+    protected fun setUpExtraRow(
+        cardView: View,
+        iconMap: EnumMap<DataType, Int>,
+        oCounter: Int?,
+    ) {
+        val infoView = cardView.findViewById<ViewGroup>(androidx.leanback.R.id.info_field)
+        val sceneExtra =
+            LayoutInflater.from(infoView.context)
+                .inflate(R.layout.scene_card_extra, infoView, true) as ViewGroup
+
+        iconMap.forEach {
+            setUpIcon(sceneExtra, it.key, it.value)
+        }
+        if (oCounter != null && oCounter > 0) {
+            setUpIcon(sceneExtra, null, oCounter)
+        }
+    }
+
+    private fun setUpIcon(
+        rootView: ViewGroup,
+        dataType: DataType?,
+        count: Int,
+    ) {
+        val textResId: Int
+        val iconResId: Int
+        when (dataType) {
+            DataType.MOVIE -> {
+                textResId = R.id.scene_movie_count
+                iconResId = R.id.scene_movie_icon
+            }
+
+            DataType.MARKER -> {
+                textResId = R.id.scene_marker_count
+                iconResId = R.id.scene_marker_icon
+            }
+
+            DataType.PERFORMER -> {
+                textResId = R.id.scene_performer_count
+                iconResId = R.id.scene_performer_icon
+            }
+
+            DataType.TAG -> {
+                textResId = R.id.scene_tag_count
+                iconResId = R.id.scene_tag_icon
+            }
+
+            DataType.SCENE -> {
+                textResId = R.id.scene_scene_count
+                iconResId = R.id.scene_scene_icon
+            }
+
+            // Workaround for O Counter
+            null -> {
+                textResId = R.id.scene_ocounter_count
+                iconResId = R.id.scene_ocounter_icon
+            }
+
+            else -> throw IllegalArgumentException()
+        }
+        val textView = rootView.findViewById<TextView>(textResId)
+        val iconView = rootView.findViewById<View>(iconResId)
+        if (count > 0) {
+            textView.text = count.toString()
+            textView.visibility = View.VISIBLE
+            iconView.visibility = View.VISIBLE
+            (textView.parent as ViewGroup).visibility = View.VISIBLE
+        }
     }
 
     companion object {
