+++ app/src/main/res/values/dimens.xml
@@ -2,5 +2,5 @@
     <integer tools:override="true" name="lb_details_description_body_max_lines">18</integer>
     <dimen name="title_bar_height">60dp</dimen>
     <dimen name="title_bar_margin">10dp</dimen>
-    <dimen name="scene_extra_spacing">5dp</dimen>
+    <dimen name="scene_extra_spacing">4dp</dimen>
 </resources>
