+++ app/src/main/res/values/strings.xml
@@ -9,6 +9,6 @@
     <string name="fa_user" translatable="false">&#xf007;</string>
     <string name="fa_location_dot" translatable="false">&#xf3c5;</string>
     <string name="fa_film" translatable="false">&#xf008;</string>
-
+    <string name="fa_circle_play" translatable="false">&#xf144;</string>
 
 </resources>
